Move things, implement ReconnectFunction
