+++ app/src/main/java/edu/drexel/lapcounter/lapcounter/backend/lapcounter/DisconnectManager.java
@@ -1,8 +1,12 @@
-package edu.drexel.lapcounter.lapcounter.backend;
+package edu.drexel.lapcounter.lapcounter.backend.lapcounter;
 
 import android.content.Context;
 import android.content.Intent;
 import android.support.v4.content.LocalBroadcastManager;
+
+import edu.drexel.lapcounter.lapcounter.backend.BLEService;
+import edu.drexel.lapcounter.lapcounter.backend.RSSIManager;
+import edu.drexel.lapcounter.lapcounter.backend.SimpleMessageReceiver;
 
 public class DisconnectManager {
     /**
